fix : Maker 구조 변경, 오류 제거
Maker 측 Service 코드 오류를 제거함.
Maker 측 Service를 메서드만 잘 볼 수 있게 인터페이스로 분리
Maker 측 dto 패키지를 따로 빼서 관리

Resolves: #issueNo
See also: #issueNo
