--- src/main/java/io/wisoft/testermatchingplatform/web/dto/response/maker/CompleteResponse.java
@@ -1,14 +1,0 @@
-package io.wisoft.testermatchingplatform.web.dto.response.maker;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-import java.util.List;
-import java.util.UUID;
-
-@Getter
-@AllArgsConstructor
-public class CompleteResponse {
-    private List<UUID> CompleteSuccessUUIDDTO;
-
-}
