+++ src/main/java/io/wisoft/testermatchingplatform/web/dto/response/maker/FindApplyResponse.java
@@ -1,5 +1,6 @@
 package io.wisoft.testermatchingplatform.web.dto.response.maker;
 
+import io.wisoft.testermatchingplatform.web.dto.response.maker.dto.ApplyDTO;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
@@ -7,7 +8,6 @@ import java.util.List;
 
 @Getter
 @AllArgsConstructor
-public class CompleteTesterListResponse {
-    List<CompleteTesterDTO> completeTesterDTOList;
-
+public class FindApplyResponse {
+    private List<ApplyDTO> applyDTOList;
 }
