+++ src/main/java/io/wisoft/testermatchingplatform/web/dto/response/maker/FindPaymentResponse.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class PaymentResponse {
+public class FindPaymentResponse {
     private long point;
     private String accountNumber;
 }
