+++ src/main/java/io/wisoft/testermatchingplatform/web/dto/response/maker/FindPerformListResponse.java
@@ -1,5 +1,6 @@
 package io.wisoft.testermatchingplatform.web.dto.response.maker;
 
+import io.wisoft.testermatchingplatform.web.dto.response.maker.dto.PerformDTO;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
@@ -7,6 +8,6 @@ import java.util.List;
 
 @Getter
 @AllArgsConstructor
-public class PerformListResponse {
+public class FindPerformListResponse {
     List<PerformDTO> performDTOList;
 }
