+++ src/main/java/io/wisoft/testermatchingplatform/web/dto/response/maker/dto/ApprovePeriodTestDTO.java
@@ -1,4 +1,4 @@
-package io.wisoft.testermatchingplatform.web.dto.response.maker;
+package io.wisoft.testermatchingplatform.web.dto.response.maker.dto;
 
 import io.wisoft.testermatchingplatform.domain.test.Test;
 import lombok.Getter;
@@ -6,22 +6,27 @@ import lombok.Getter;
 import java.util.UUID;
 
 @Getter
-public class CompletePeriodTestDTO {
+public class ApprovePeriodTestDTO {
     private UUID id;
     private String title;
     private String makerNickname;
     private String company;
     private int reward;
-    private String state;
+    private int participantCapacity;
+    private int apply;
     private String symbolImageRoot;
 
-    public static CompletePeriodTestDTO fromEntity(Test test, String state) {
-        CompletePeriodTestDTO dto = new CompletePeriodTestDTO();
+    private String state;
+
+    public static ApprovePeriodTestDTO fromEntity(Test test, int applyCount, String state) {
+        ApprovePeriodTestDTO dto = new ApprovePeriodTestDTO();
         dto.id = test.getId();
         dto.title = test.getTitle();
         dto.makerNickname = test.getMaker().getNickname();
         dto.company = test.getMaker().getCompany();
         dto.reward =test.getReward();
+        dto.apply = applyCount;
+        dto.participantCapacity = test.getParticipantCapacity();
         dto.symbolImageRoot = test.getSymbolImageRoot();
         dto.state = state;
