+++ src/main/java/io/wisoft/testermatchingplatform/web/dto/response/maker/dto/CompleteTesterDTO.java
@@ -1,7 +1,6 @@
-package io.wisoft.testermatchingplatform.web.dto.response.maker;
+package io.wisoft.testermatchingplatform.web.dto.response.maker.dto;
 
 import io.wisoft.testermatchingplatform.domain.tester.Tester;
-import lombok.Data;
 import lombok.Getter;
 
 import java.util.UUID;
