+++ src/main/java/io/wisoft/testermatchingplatform/web/dto/response/maker/findTestsResponse.java
@@ -1,6 +1,9 @@
 package io.wisoft.testermatchingplatform.web.dto.response.maker;
 
-import io.wisoft.testermatchingplatform.domain.test.Test;
+import io.wisoft.testermatchingplatform.web.dto.response.maker.dto.ApplyPeriodTestDTO;
+import io.wisoft.testermatchingplatform.web.dto.response.maker.dto.ApprovePeriodTestDTO;
+import io.wisoft.testermatchingplatform.web.dto.response.maker.dto.CompletePeriodTestDTO;
+import io.wisoft.testermatchingplatform.web.dto.response.maker.dto.ProgressPeriodTestDTO;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
@@ -8,7 +11,7 @@ import java.util.List;
 
 @Getter
 @AllArgsConstructor
-public class TestsFromMakerResponse {
+public class findTestsResponse {
     private List<ApplyPeriodTestDTO> applyPeriodTestList;
     private List<ApprovePeriodTestDTO> approvePeriodTestList;
     private List<ProgressPeriodTestDTO> progressPeriodTestList;
